requirements.txt
asgiref==3.5.2
Django==3.2.13
django-crispy-forms==1.14.0
django-tinymce==3.5.0
Pillow==9.0.1
pytz==2022.4
sqlparse==0.4.3
